+++ .gitignore
@@ -1,9 +1,26 @@
-*.pyc
+*.py[cod]
+*$py.class
 *.egg-info
+*.log
+*.pot
 .DS_Store
-.idea/
-.tox/
+.coverage/
 .eggs/
-dist/
+.idea/
+.project/
+.pydevproject/
+.settings/
+.tox/
+__pycache__/
 build/
-node_modules/
+dist/
+env/
+
+/~
+/node_modules
+.sass-cache
+*.css.map
+npm-debug.log
+package-lock.json
+
+local.sqlite
