+++ .travis.yml
@@ -6,10 +6,12 @@ matrix:
   include:
     - python: 3.5
       env: TOX_ENV='flake8'
-    # Django 1.11
+    - python: 3.5
+      env: TOX_ENV='isort'
+      # Django 1.11
     - python: 2.7
       env: DJANGO='dj111' CMS='cms34'
-    - python: 3.5
+    - python: 3.4
       env: DJANGO='dj111' CMS='cms35'
     - python: 3.5
       env: DJANGO='dj111' CMS='cms36'
@@ -17,13 +19,15 @@ matrix:
     - python: 3.5
       env: DJANGO='dj20' CMS='cms36'
     # Django 2.1
-    - python: 3.5
+    - python: 3.6
       env: DJANGO='dj21' CMS='cms36'
 
 install:
-  - pip install tox coverage
+  - pip install coverage isort tox
   - "if [[ $TRAVIS_PYTHON_VERSION == '2.7' ]]; then export PY_VER=py27; fi"
+  - "if [[ $TRAVIS_PYTHON_VERSION == '3.4' ]]; then export PY_VER=py34; fi"
   - "if [[ $TRAVIS_PYTHON_VERSION == '3.5' ]]; then export PY_VER=py35; fi"
+  - "if [[ $TRAVIS_PYTHON_VERSION == '3.6' ]]; then export PY_VER=py36; fi"
   - "if [[ ${DJANGO}z != 'z' ]]; then export TOX_ENV=$PY_VER-$DJANGO-$CMS; fi"
 
 script:
