+++ CHANGELOG.rst
@@ -2,6 +2,15 @@
 Changelog
 =========
 
+
+1.2.1 (unreleased)
+==================
+
+* Extended test matrix
+* Added isort and adapted imports
+* Adapted code base to align with other supported addons
+
+
 1.2.0 (2018-12-12)
 ==================
