+++ MANIFEST.in
@@ -3,4 +3,4 @@ include README.rst
 recursive-include djangocms_icon/locale *
 recursive-include djangocms_icon/static *
 recursive-include djangocms_icon/templates *
-ecursive-exclude * *.pyc
+recursive-exclude * *.py[co]
