+++ README.rst
@@ -2,7 +2,6 @@
 django CMS Icon
 ===============
 
-
 |pypi| |build| |coverage|
 
 **django CMS Icon** is a plugin for `django CMS <http://django-cms.org>`_
@@ -23,6 +22,10 @@ feedback in the form of issues and pull requests. Before submitting your
 pull request, please review our `contribution guidelines
 <http://docs.django-cms.org/en/latest/contributing/index.html>`_.
 
+We're grateful to all contributors who have helped create and maintain this package.
+Contributors are listed at the `contributors <https://github.com/divio/djangocms-icon/graphs/contributors>`_
+section.
+
 One of the easiest contributions you can make is helping to translate this addon on
 `Transifex <https://www.transifex.com/projects/p/djangocms-icon/>`_.
 
@@ -33,8 +36,7 @@ Documentation
 See ``REQUIREMENTS`` in the `setup.py <https://github.com/divio/djangocms-icon/blob/master/setup.py>`_
 file for additional dependencies:
 
-* Python 2.7, 3.4 or higher
-* Django 1.11 or higher
+|python| |django| |djangocms|
 
 
 Installation
@@ -177,3 +179,10 @@ You can run tests by executing::
     :target: https://travis-ci.org/divio/djangocms-icon
 .. |coverage| image:: https://codecov.io/gh/divio/djangocms-icon/branch/master/graph/badge.svg
     :target: https://codecov.io/gh/divio/djangocms-icon
+
+.. |python| image:: https://img.shields.io/badge/python-2.7%20%7C%203.4+-blue.svg
+    :target: https://pypi.org/project/djangocms-icon/
+.. |django| image:: https://img.shields.io/badge/django-1.11%20%7C%202.0%20%7C%202.1-blue.svg
+    :target: https://www.djangoproject.com/
+.. |djangocms| image:: https://img.shields.io/badge/django%20CMS-3.4%2B-blue.svg
+    :target: https://www.django-cms.org/
