+++ setup.py
@@ -4,6 +4,7 @@ from setuptools import find_packages, setup
 
 from djangocms_icon import __version__
 
+
 REQUIREMENTS = [
     'django-cms>=3.4.5',
     'djangocms-attributes-field>=0.4.0',
@@ -13,20 +14,24 @@ REQUIREMENTS = [
 CLASSIFIERS = [
     'Development Status :: 5 - Production/Stable',
     'Environment :: Web Environment',
-    'Framework :: Django',
     'Intended Audience :: Developers',
     'License :: OSI Approved :: BSD License',
     'Operating System :: OS Independent',
+    'Framework :: Django',
+    'Framework :: Django :: 1.11',
+    'Framework :: Django :: 2.0',
+    'Framework :: Django :: 2.1',
     'Programming Language :: Python',
     'Programming Language :: Python :: 2',
     'Programming Language :: Python :: 2.7',
     'Programming Language :: Python :: 3',
     'Programming Language :: Python :: 3.4',
     'Programming Language :: Python :: 3.5',
+    'Programming Language :: Python :: 3.6',
     'Topic :: Internet :: WWW/HTTP',
     'Topic :: Internet :: WWW/HTTP :: Dynamic Content',
-    'Topic :: Software Development :: Libraries :: Application Frameworks',
-    'Topic :: Software Development :: Libraries :: Python Modules',
+    'Topic :: Software Development',
+    'Topic :: Software Development :: Libraries',
 ]
