+++ tests/settings.py
@@ -1,6 +1,5 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
-
 HELPER_SETTINGS = {
     'INSTALLED_APPS': [],
     'CMS_LANGUAGES': {
