+++ tox.ini
@@ -1,6 +1,7 @@
 [tox]
 envlist =
     flake8
+    isort
     py{27,34,35,36}-dj111-cms{34,35,36}
     py{34,35,36}-dj20-cms36
     py{35,36}-dj21-cms36
@@ -23,5 +24,30 @@ commands =
     {env:COMMAND:coverage} report
 
 [flake8]
-max-line-length = 120
-exclude = */docs/*,*/migrations/*
+max-line-length = 119
+exclude =
+    *.egg-info,
+    .eggs,
+    .git,
+    .settings,
+    .tox,
+    build,
+    data,
+    dist,
+    docs,
+    *migrations*,
+    requirements,
+    tmp
+
+[isort]
+line_length = 79
+skip = manage.py, *migrations*, .tox, .eggs, data
+include_trailing_comma = true
+multi_line_output = 5
+not_skip = __init__.py
+lines_after_imports = 2
+default_section = THIRDPARTY
+sections = FUTURE, STDLIB, DJANGO, CMS, THIRDPARTY, FIRSTPARTY, LIB, LOCALFOLDER
+known_first_party = djangocms_icon
+known_cms = cms, menus
+known_django = django
